pull logic for creating composition out from createClient
